Fix/Change: Limit to org-mode in global instead of local ushin mode
Previously the mode got "enabled" in all buffers, it just didn't have
any effect in non-org buffers.  As a side-effect it is now possible to
properly enable the mode in non-org buffers, which may occasionally be
useful, or at least is better than to always enable a dysfunctional
variant of the mode.